Customer Churn Analysis using Python
Customer churn analysis is a process of analyzing customer behavior over time to identify too frequently those customers who are at risk of leaving your business. This analysis can help you identify which actions you need to take to reduce churn and retain customers. In this article, we will be discussing how to use Python to perform this analysis.